# frontend/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "BackgroundJobResponse": {
        "properties": {
          "failure_kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Kind"
          },
          "failure_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Message"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "job_type": {
            "title": "Job Type",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "progress_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress Message"
          },
          "requested_by": {
            "title": "Requested By",
            "type": "string"
          },
          "retryable": {
            "default": false,
            "title": "Retryable",
            "type": "boolean"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": ["job_id", "session_id", "state", "kind", "job_type", "title", "requested_by"],
        "title": "BackgroundJobResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PageInfoResponse": {
        "properties": {
          "cursor": {
            "title": "Cursor",
            "type": "integer"
          },
          "has_more": {
            "title": "Has More",
            "type": "boolean"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "returned_count": {
            "title": "Returned Count",
            "type": "integer"
          }
        },
        "required": ["cursor", "limit", "next_cursor", "has_more", "returned_count"],
        "title": "PageInfoResponse",
        "type": "object"
      },
...
      "RepositoryIndexEntryDetailResponse": {
        "properties": {
          "entry": {
            "$ref": "#/components/schemas/RepositoryIndexEntryResponse"
          }
        },
        "required": ["entry"],
        "title": "RepositoryIndexEntryDetailResponse",
        "type": "object"
      },
      "RepositoryIndexEntryResponse": {
        "properties": {
          "entry_id": {
            "title": "Entry Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path"
          },
          "provenance": {
            "items": {
              "$ref": "#/components/schemas/RepositoryIndexProvenanceResponse"
            },
            "title": "Provenance",
            "type": "array"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Symbol"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": ["entry_id", "kind", "name", "provenance", "tags", "updated_at"],
        "title": "RepositoryIndexEntryResponse",
        "type": "object"
      },
      "RepositoryIndexProvenanceResponse": {
        "properties": {
          "content_sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Sha256"
          },
          "line_end": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Line End"
          },
          "line_start": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Line Start"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path"
          },
          "source_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Label"
          },
          "source_type": {
            "title": "Source Type",
            "type": "string"
          },
          "tool_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool Name"
          }
        },
        "required": ["source_type"],
        "title": "RepositoryIndexProvenanceResponse",
        "type": "object"
      },
      "RepositoryIndexRebuildRequest": {
        "properties": {
          "background": {
            "default": true,
            "title": "Background",
            "type": "boolean"
          },
          "requested_by": {
            "default": "operator",
            "title": "Requested By",
            "type": "string"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          }
        },
        "title": "RepositoryIndexRebuildRequest",
        "type": "object"
      },
      "RepositoryIndexRebuildResponse": {
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "index": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RepositoryIndexStatusResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "job": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BackgroundJobResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "mode": {
            "title": "Mode",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": ["mode", "status"],
        "title": "RepositoryIndexRebuildResponse",
        "type": "object"
      },
      "RepositoryIndexSearchPageResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RepositoryIndexEntryResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "$ref": "#/components/schemas/PageInfoResponse"
          },
          "query": {
            "title": "Query",
            "type": "string"
          }
        },
        "required": ["query", "page", "items"],
        "title": "RepositoryIndexSearchPageResponse",
        "type": "object"
      },
      "RepositoryIndexStatusResponse": {
        "properties": {
          "builder_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Builder Version"
          },
          "built_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Built At"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "entry_count": {
            "title": "Entry Count",
            "type": "integer"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "schema_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schema Version"
          },
          "source_digest": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Digest"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": ["status", "path", "entry_count"],
        "title": "RepositoryIndexStatusResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": ["loc", "msg", "type"],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Glassbox",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/repo/index/entries/{entry_id}": {
      "get": {
        "description": "Return one repository index entry by stable ID.",
        "operationId": "get_repository_index_entry_detail_repo_index_entries__entry_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "title": "Entry Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepositoryIndexEntryDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Repository Index Entry Detail",
        "tags": ["repo"]
      }
    },
    "/repo/index/rebuild": {
      "post": {
        "description": "Refresh repository intelligence directly or queue a daemon job.",
        "operationId": "rebuild_repository_index_repo_index_rebuild_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RepositoryIndexRebuildRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepositoryIndexRebuildResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rebuild Repository Index",
        "tags": ["repo"]
      }
    },
    "/repo/index/search": {
      "get": {
        "description": "Search repository index entries by text.",
        "operationId": "search_repository_index_entries_repo_index_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Query",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepositoryIndexSearchPageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Repository Index Entries",
        "tags": ["repo"]
      }
    },
...
  }
}